+++ NEWS.md
@@ -3,6 +3,7 @@ PENDING: Dataiter 0.50
 
 * `ListOfDicts.drop_na`: New method
 * `ListOfDicts.keys`: New method
+* `ListOfDicts.print_memory_use`: New method
 
 2023-11-08: Dataiter 0.49
 =========================